+++ src/tui/dummy.go
@@ -33,8 +33,7 @@ func (r *FullscreenRenderer) Init()                              {}
 func (r *FullscreenRenderer) Resize(maxHeightFunc func(int) int) {}
 func (r *FullscreenRenderer) Pause(bool)                         {}
 func (r *FullscreenRenderer) Resume(bool, bool)                  {}
-func (r *FullscreenRenderer) PassThrough(int, int, string)       {}
-func (r *FullscreenRenderer) Sync(bool)                          {}
+func (r *FullscreenRenderer) PassThrough(string)                 {}
 func (r *FullscreenRenderer) Clear()                             {}
 func (r *FullscreenRenderer) NeedScrollbarRedraw() bool          { return false }
 func (r *FullscreenRenderer) Refresh()                           {}