+++ CHANGELOG.md
@@ -3,19 +3,18 @@ CHANGELOG
 
 0.44.0
 ------
-- (Experimental) Sixel image support in preview window
+- (Experimental) Sixel image support in preview window (not available on Windows)
     - [bin/fzf-preview.sh](bin/fzf-preview.sh) is added to demonstrate how to
       display an image using Kitty image protocol or Sixel. You can use it
       like so:
       ```sh
       fzf --preview='fzf-preview.sh {}'
       ```
-- (Experimental) iTerm2 inline image protocol support in preview window
+- (Experimental) iTerm2 inline image protocol support in preview window (not available on Windows)
   ```sh
   # Using https://iterm2.com/utilities/imgcat
   fzf --preview 'imgcat -W $FZF_PREVIEW_COLUMNS -H $FZF_PREVIEW_LINES {}'
   ```
-- (Experimental) Sixel, Kitty, and iTerm2 image support now also available on Windows
 - HTTP server can be configured to accept remote connections
   ```sh
   # FZF_API_KEY is required for a non-localhost listen address